Release 0.2.0-beta2

+++ home/rules/comparison/min_count.py
@@ -1,5 +1,5 @@
 #
-# Copyright © 2021 DQO.ai (support@dqo.ai)
+# Copyright © 2021 DQOps (support@dqops.com)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@ from typing import Sequence
 
 # rule specific parameters object, contains values received from the quality check threshold configuration
 class MinCountRuleParametersSpec:
-    min_count: float
+    min_count: int
 
 
 class HistoricDataPoint:
